+++ packages/commitlint-plugin/src/rule.ts
@@ -1,36 +1,44 @@
 import type { Rule } from '@commitlint/types';
+import emojiRegex from 'emoji-regex';
+
 import { gitmojiCodes, gitmojiUnicode } from './gitmojiCode';
+
+const gitjimojiCodeStr = `:\\w*:`;
+const gitmojiUnicodeStr = gitmojiUnicode.filter((i) => i).join('|');
+const emojiStr = emojiRegex().source;
 
 const emoji: Rule = (parsed) => {
   const { raw } = parsed;
 
   // code regex test url: https://regex101.com/r/fSdOvB/1
-  const regex = /^(:\w*:)\s.*/gm;
-  // unicode regex test url: https://regex101.com/r/OTMgWL/2
-  const unicodeRegex =
-    /(\ud83c[\udf00-\udfff]|\ud83d[\udc00-\ude4f\ude80-\udeff]|[\u2600-\u2B55])\s.*/gm;
-
-  const result = regex.exec(raw);
-  const unicodeResult = unicodeRegex.exec(raw);
+  const gitmojiCodeResult = new RegExp(`(${gitjimojiCodeStr})\\s.*`, 'gm').exec(raw);
+  // unicode regex test url: https://regex101.com/r/shBTBg/2
+  const gitmojiUnicodeResult = new RegExp(`(${gitmojiUnicodeStr})\\s.*`, 'gm').exec(raw);
+  const emojiResult = new RegExp(`(${emojiStr})\\s.*`, 'gm').exec(raw);
 
   let pass;
   let errorMsg = 'passed';
 
   // if gitmoji code is valid
-  if (result) {
-    const emojiCode = result[1];
+  if (gitmojiCodeResult) {
+    const emojiCode = gitmojiCodeResult[1];
     pass = gitmojiCodes.includes(emojiCode);
     if (!pass) {
       errorMsg = `${emojiCode} is not in the correct gitmoji list, please check the emoji code on https://gitmoji.dev/.`;
     }
-  } else if (unicodeResult) {
-    const unicode = unicodeResult[1];
+  }
+  // if gitmoji unicode is valid
+  else if (gitmojiUnicodeResult) {
+    const unicode = gitmojiUnicodeResult[1];
 
     pass = gitmojiUnicode.includes(unicode);
+  }
+  // is emoji,but isn't included in gitmoji list
+  else if (emojiResult) {
+    const unicode = emojiResult[1];
 
-    if (!pass) {
-      errorMsg = `${unicode} is not in the correct gitmoji list, please check the emoji code on https://gitmoji.dev/.`;
-    }
+    pass = false;
+    errorMsg = `${unicode} is not in the correct gitmoji list, please check the emoji code on https://gitmoji.dev/.`;
   } else {
     // if don't has gitmoji code or emoji unicode
     pass = false;